feat: add inline keyboard for a message, change the logic of subscription checking (check 2 chanells)

+++ handlers/check_subscription.py
@@ -16,10 +16,10 @@ from utils import messages_for_users
 router = Router()
 logger = get_logger(__name__)
 
-CHANNEL_TO_CHECK = settings.CHANNEL_TO_CHECK
+CHANNEL1_TO_CHECK = settings.CHANNEL1_TO_CHECK
+CHANNEL2_TO_CHECK = settings.CHANNEL2_TO_CHECK
 
 
-# TODO: make the function check two tg channels
 @router.callback_query(
     StateFilter(FSMRegisterUser.subscribe_channels), F.data.in_(['subscribed'])
 )
@@ -28,13 +28,18 @@ async def check_tg_channels_subscription(
 ):
     user_id = callback.from_user.id
     try:
-        member = await bot.get_chat_member(CHANNEL_TO_CHECK, user_id)
-        if member.status == 'member':
+        channel_1 = await bot.get_chat_member(CHANNEL1_TO_CHECK, user_id)
+        channel_2 = await bot.get_chat_member(CHANNEL2_TO_CHECK, user_id)
+
+        if channel_1.status == 'member' and channel_2.status == 'member':
             await callback.message.delete()
             await callback.message.answer(messages_for_users.WAIT_FOR_SCREENSHOT_MESSAGE)
             await state.set_state(FSMRegisterUser.send_valid_pass)
         else:
-            await callback.message.answer(messages_for_users.NOT_SUBSCRIBED_MESSAGE)
+            markup = prepare_inline_keyboard()
+            await callback.message.answer(
+                messages_for_users.NOT_SUBSCRIBED_MESSAGE, reply_markup=markup
+            )
     except Exception as e:
         logger.error(f'Failed to check subscription: {e}')
         await callback.answer(messages_for_users.GENERIC_ERROR_MESSAGE)
@@ -42,12 +47,17 @@ async def check_tg_channels_subscription(
 
 @router.message(StateFilter(FSMRegisterUser.subscribe_channels))
 async def warning_not_subscribed_press(message: Message):
+    markup = prepare_inline_keyboard()
+    await message.answer(
+        messages_for_users.BUTTON_IS_NOT_PRESSED_MESSAGE, reply_markup=markup
+    )
+
+
+def prepare_inline_keyboard() -> InlineKeyboardMarkup:
     subscribed_button = InlineKeyboardButton(
         text='Я подписался(ась)',
         callback_data='subscribed'
     )
     keyboard: list[list[InlineKeyboardButton]] = [[subscribed_button]]
     markup = InlineKeyboardMarkup(inline_keyboard=keyboard)
-    await message.answer(
-        messages_for_users.BUTTON_IS_NOT_PRESSED_MESSAGE, reply_markup=markup
-    )
+    return markup
